# docs/openapi/schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LecheFacil Backend",
    "description": "Multi-tenant API for LecheFacil MVP",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/animals/": {
      "get": {
        "tags": [
          "animals"
        ],
        "summary": "List Animals Endpoint",
        "operationId": "list_animals_endpoint_api_v1_animals__get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalsListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "animals"
        ],
        "summary": "Create Animal Endpoint",
        "operationId": "create_animal_endpoint_api_v1_animals__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnimalCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reports/production": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Generate Production Report",
        "description": "Generate production report",
        "operationId": "generate_production_report_api_v1_reports_production_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/financial": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Generate Financial Report",
        "description": "Generate financial report",
        "operationId": "generate_financial_report_api_v1_reports_financial_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/animals": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Generate Animals Report",
        "description": "Generate animals report",
        "operationId": "generate_animals_report_api_v1_reports_animals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reports/export-all": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Export All Reports",
        "description": "Export all reports",
        "operationId": "export_all_reports_api_v1_reports_export_all_post",
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date To"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "pdf",
                "json"
              ],
              "type": "string",
              "default": "pdf",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReportResponse"
                  },
                  "title": "Response Export All Reports Api V1 Reports Export All Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnimalCreate": {
        "properties": {
          "tag": {
            "type": "string",
            "title": "Tag"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "breed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Breed"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "lot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lot"
          },
          "status": {
            "$ref": "#/components/schemas/AnimalStatus",
            "default": "active"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          }
        },
        "type": "object",
        "required": [
          "tag"
        ],
        "title": "AnimalCreate"
      },
      "AnimalResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "tag": {
            "type": "string",
            "title": "Tag"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "breed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Breed"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "lot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lot"
          },
          "status": {
            "$ref": "#/components/schemas/AnimalStatus"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "primary_photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Photo Url"
          },
          "photos_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photos Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "tag",
          "name",
          "breed",
          "birth_date",
          "lot",
          "status",
          "photo_url",
          "created_at",
          "updated_at",
          "version"
        ],
        "title": "AnimalResponse"
      },
      "AnimalStatus": {
        "type": "string",
        "enum": [
          "active",
          "sold",
          "dead",
          "culled"
        ],
        "title": "AnimalStatus"
      },
...
      "AnimalsListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnimalResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "AnimalsListResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportFilters": {
        "properties": {
          "animal_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Animal Ids"
          },
          "buyer_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buyer Ids"
          },
          "include_inactive": {
            "type": "boolean",
            "title": "Include Inactive",
            "default": false
          }
        },
        "type": "object",
        "title": "ReportFilters"
      },
...
      "ReportRequest": {
        "properties": {
          "date_from": {
            "type": "string",
            "format": "date",
            "title": "Date From"
          },
          "date_to": {
            "type": "string",
            "format": "date",
            "title": "Date To"
          },
          "period": {
            "type": "string",
            "enum": [
              "daily",
              "weekly",
              "monthly"
            ],
            "title": "Period",
            "default": "daily"
          },
          "format": {
            "type": "string",
            "enum": [
              "pdf",
              "json"
            ],
            "title": "Format",
            "default": "pdf"
          },
          "filters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportFilters"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "date_from",
          "date_to"
        ],
        "title": "ReportRequest"
      },
      "ReportResponse": {
        "properties": {
          "report_id": {
            "type": "string",
            "title": "Report Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "file_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Name"
          }
        },
        "type": "object",
        "required": [
          "report_id",
          "title",
          "generated_at",
          "format"
        ],
        "title": "ReportResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}